Move error summary service
